Move NetBox in separate file, add cache
